[#116] keep selection if metric could be saved
The table was being reloaded upon the goal
being succesfully updated.
Reloading the table canceled the selected row, which is what
was being used to display the user's selection.

Considering the VC is being dismissed (popped) immediately
afterward, that can be skipped altogether instead of tracking
the selected row (metric) to mark it as such again when
reloading the table.