fix #151 Prepare today widget goal cell for reuse
The views (labels, stepper, button, imageview) are
initialized when the class is constructed.
Reusing the cell does not create these subviews anew
but uses the already constructed view, including its subviews.

Here the subviews are cleared of their constraints
since they will all be set again according to the
most recent goal assigned to the cell.