Configure Cell for re-use
datapoint cell was being reused but not cleared out/prepared for such.
The tableview on the goalVC would only set the text of the cell
for as many rows as the goal currently had of recent_data.
In the case of say 5 datapoints in recent_data, followed by
in-app deletion of one of the earlier datapoints, goalVC
would update the text of the cells for four by not the fifth,
which would still be showing data from earlier.

Here we use prepareForReuse to set the cell back to its initial state
(clear out the text being displayed).

Fixes #166