Fix navigation and status bar colors when built for iOS 15
Due to changes in a recent iOS version, navigation controllers now default to
having a transparent background, which meant the intended black background
was not shown.

Here we move UI configuration into the Storyboard file, and update so that
the background is always opaque, fixing the rendering.

Manual Testing:
Viewed in the simulator on the gallery and goals page, in light and dark
modes