Remove unused beeminder API fields
Fixes #292

The iOS app was recently broken by removal of the `panic` field from the
beeminder API. This wasn't used for anything, but the deserialization code
assumed it existed, and thus crashed.

Here we remove `panic`, along with other properties which are easy to prove
the app does not use.

Test Plan:
TODO