Make Shortcuts behavior in development match production.
Previously Intents for Shortcuts were crashing in production but not
in dev builds. This was due to a setting deep in build settings which
meant the code to handle intents was not being installed at all during
development, but was for App Store installs. This meant crashes in the
code could not be debugged locally.

Here we change the flag to also install this code while debugging.

This is the first step towards addressing #293 but does not yet fix the
bug.

Test Plan:
Ran a shortcut in simulator and observed the same error in console as
App Store builds have been producing