Use the complete width of the device when rendering gallery view (#291)
Previously the gallery view was limited to 320px. This added large margins
and meant descriptions were often unnecessarily truncated. Here we change the
width calculation logic so items are always at least 320px, but will be wider
if there is space. We continue to favor showing multiple columns over making
items wider.

Test Plan:
Viewed in both portrait and landscape in the simulator on the following
devices:
* iPhone 8
* iPhone 13 Pro
* iPhone 13 Max