Remove arm64 as an excluded architecture for some targets
Previously the project was configured to exclude arm64 for some targets - possibly so
that tests did not get built for iOS device targets. However, since Apple has released
M1 macs it is possible for macs to also be arm64, and this setting means tests could
not be run. Amongst other things this broke `fastlane ios test`. We address this by removing
this filter.

Test Plan:
Ran `fastline ios test`