Fix the AddData Intent (#303)
Previous commits had begun to implement an intent to add data points to beeminder. This follows up on that work, adding the necessary extension routing and parameter checking to make intents support work without launching the app. This should be seen as a minimum viable implementation, there remain many possible improvements to e.g. error handling, and suggesting allowed goal names.

As datapoint submission is now handled the the extension this removes the code which submitted them as part of app startup.

Fixes #293