Fix multiple settings icons being overlaid
Previously we were not properly removing settings icons when cells
in the settings table are re-used. Here we use a shared UIImageView
which has its image updated, avoiding multiple being shown.

Fixes: #306

Test Plan:
Navigate to settings, back out, and then back to settings again.
Observe that icons are not mis-rendered.