Do not allow description to hide todayta indicator
Previously with a very long description, the autolayout system could decide
to hide the todayta tick to show more of the description. Here we make the
view holding the tick impossible to shrink so it always appears.

Fixes #305

Test Plan:
Made a goal description very long, observed before this fix the todayta is
hidden, and after the change it is shown.