Run tests through github actions on PR (#316)
This project was previously configured to use semaphore for unit tests. However
lack of credits meant a slow queue. Github actions now has the capability to run
these tests, and is more consistent with what is used elsewhere at beeminder.

This causes github to run the test suite on PRs.

Note: The test suite is currently broken. That is not addressed by this PR.