Fix display of & in datapoint list
For some reason the canonical representation of data points html-encodes `&`. To
avoid this we instead reconstruct the text from the individual fields. In the future
we could go further and use columns, but for now we just re-assemble the original
text output.

Fixes #283

Test Plan:
Added & to a datapoint and viewed the goal in the simulator