Isolate HealthKit interactions from the rest of the codebase
This PR takes two steps towards isolating HealthKit logic from the rest of the codebase
1. It removes logic from the JSONGoal class, and instead moves it to a HealthKitConnection class
2. It moves all health kit data model operations to behind HealthStoreManager

This reduction in surface area should make it much simpler to refactor the health kit code.

Test Plan:
* Run on device and observe sync with health kit buttons work
* Also add significant logging and observe it is run as expected

Note that this code cannot be correctly tested on simulator as it does not have data available.