Use HKObserverQuery for all metric types
Previously we were using HKObserverQuery for some metric types, and
HKStatisticsCollectionQuery for other types. The latter does have a
mechanism to deliver updates, but it is intended to only be called back
while the app is running, and doesn't get called in the background. This means
these metrics were not regularly being updated.

Instead here we use HKObserverQuery everywhere, and run an appropriate
follow-up query to fetch data when it reports a change.