Refactor VersionManager to be async
This changes the VersionManager (which compares the current app version to
the latest available and supported) to expose an async rather than a callback
based interface. It also somewhat simplifies the interface, with the manager
now responsible for returning an enum, rather than clients having to do version
comparisons.

Test Plan:
* Ran the app in the simulator
* Patched the VersionManager to return different values and checked the UI rendered correctly