Only prompt to upgrade when a new version has been out for 10 days
The app checks the current app store version and prompts to upgrade if the
current version is older. However, because we roll out incrementally over
7 days this means users will often get prompted (when we would rather they instead
silently auto-update). To avoid this we now only prompt when a release has been
out for 10 days.

Note we don't have a full release history, so if a user is on a version several
versions out of date, they may incorrectly not be shown a message for a few days
shortly after the newest version is released. This doesn't seem a big problem.

Fixes #328

Test Plan:
Overrode the current version and current date and checked the message appeared/did not appear
appropriately