Upgrade Alamofire and related packages (#342)
We were using a relatively old version of Alamofire. For future changes we want to be able to use async support, so upgrade to the current version. Minimal updates have been made to address a few backwards incompatible changes.

Test Plan:
Ran the app, loaded a goal, posted a data point.