Update colon keyboard accessory to match keyboard look
This restyles the ":" button added above the keyboard to much more closely
resemble other keyboard buttons. It is based on the appearance of iOS 16 so
may look less similar on other iOS versions. Unfortunately there are not
common controls provided for this, so we match by carefully looking at the
pixels.

Test Plan:
* Check the look in light and dark mode
* Check transitioning between the two modes works in simulator