Do not show the colon button on iPads
On iPads there is not a numeric-only keyboard, rather there is a combined
numbers-and-symbols full keyboard. As this already includes a ":" there is
no need for us to sacrifice screen space adding an extra button. So detect
when running on iPad and hide the accessory.

May help with #335

Test Plan:
* Viewed on iPad in simulator and verified button isn't shown