Convert CurrentUserManger to use async
This continues the conversion of data model code to async, with the boundary
between sync and async code now largely happening in UI handlers. Unfortunately
we still have to be a little careful to only call into NotificationCenter
from the main thread.

Test Plan:
Launched app in simulator and clicked around, including logging out and in