Keyboard should not hide datapoint submit button
When the virtual keyboard appears, we scroll to make sure text inputs are
always visible. When adding data points this is not enough - we also want the
submit button to remain visible. We do this by adding extra keyboard padding
to the relevant controls.

Fixes #335

Test Plan:
In ipad simulator, focused the relevant controls and observed the submit button was visible