Expose Datapoints as a structured type, rather than JSON (#353)
Previously while we had a Goal type, the UI code was responsible for interpreting raw JSON for data points. Here we add a DataPoint type, and make the model layer responsible for translating the JSON into this type.

Test Plan:
Confirmed the app loads data, and can edit data points.