Extract the datapoint table view into a controller (#354)
The goal view page includes a table of recent data points. Here we extract that table into a separate (nested) view controller which takes a set of data points as input. This will allow us to re-use this to show a preview of healthkit data points.

Test Plan:
Viewed the goal view for various goals
Confirmed `hhmmformat` was preserved
Confirmed tapping through to the edit screen works