Replace background updates with new framework (#363)
In iOS13 Apple rolled out a new framework for processing background updates.
Now we only support iOS14+ we can migrate to this framework. We expect the behavior
to be the same, but will stop producing warnings at compile time.

Test Plan:
Run on my device and verify logging shows the updates are running.