Goal syncing should support arbitrary date ranges
Previously when syncing healthkit datapoints to the server data, we would always fetch
exactly 7 server points. This would go badly if there were multiple points for a day
on the server, and didn't support adding features where we synced more history. Here
we instead estimate the number of points to fetch from the server, and fetch more until
we have confirmed we have enough.

Tests:
* Changed code temporarily to sync 30 points and verified it worked
* Created a large number of points for a day, and verified in the debugger they were all fetched once.