Re-use goal objects where possible (#368)
We often pass around goal objects to other classes which keep references. Previously we would have many copies of the same goal, which could allow drift, or copies to get out of date. Now we instead maintain a single copy of each goal, which we can update in place.

Testing:
Loaded and clicked around the app.