Avoid no-op datapoint updates (#372)
Previously we would always send new data point values to beeminder, even if the
old and new values were the same. This could lead to a large number of unnecessary
API calls, slowing things down and increasing server load. Now instead we check if
the existing value matches (with a small allowable epsilon) and only update on differences

Testing:
Ran the code on device and observed fewer (but some) updates