Reduce wall clock time spend during background updates
iOS applies some complex logic to decide how often to allow apps to run background updates. The
rules are opaque, but potentially include the app's energy use. To try to maximize the chance of
beeminder waking up to update goals and badges, here we change background processesing to complete
as quickly as possible

For background updates, do work as quickly as possible
For healthkit observers report as soon as we are done, even on failure.

Testing:
Observed in the debugger than background updates were still happening, and completing much
quicker than before.