Sort goals correctly (#382)
Beeminder has an intended gallery default sort order of

    min(losedate, goaldate), -pledge, slug

This is not implemented universally, but is the intended behavior. Here we update the app
to sort by this fully determined sort, rather than the partial ordering it was using before.
This also guarantees the sort is stable.

Test Plan:
Observed my gallery was sorted appropriately.