Add a screen to view and save logs (#384)
When debugging issues with background notifications it has been a pain
to use sysdiagnose to extract and find the logs. To simplify the process
the app can now display its own logs. Also make it possible to share the
logs via normal mechanisms, for example to share with support.

Testing:
Manually tested viewing and sharing logs
Added unit tests for the log reading/saving code