Access token is always supplied by the RequestManager (#395)
The BeeSwift RequestManager adds authentication information to all
requests. However, in the past this would not work in some scenarios,
for example when code is running from extensions. To work around this some
actions would also explicitly add the access token to the URL.

RequestManager should now always add the credentials (as they are stored
in a location shared with the app group) and so this should no longer be
necessary. Remove all locations in the code base where we manually passed through
credentials to simplify the code, and reduce the chance of them ending up in logs.

Testing:
* [x] Verify you can still add and refresh in the today extension
* [x] Verify you can still add data through shortcuts
* [x] Verify modified endpoints in the main beeswift app still work