Do not crash today widget when goals unavailable (#398)
The today widget previously assumed that the app had made goals
available, and
would crash if it was empty. This could happen if the user had never
logged in,
or never fetched goals. Now instead just show an empty UI. We could
ideally show
an error state, but given today widgets are deprecated we'll accept not
crashing.

To test this more easily also clear the today widget on logout. This
should fix
#236.

Testing:
Logged out and verified the today widget did not crash in the xcode
debugger