Workout minutes metric support (#401)
Adds a new metric for workout minutes, similar to other Apple Health
metrics. Implementation is slightly different due to the different
health
type.

Arguably this would be more elegantly represented in hours (using hhmm
format),
but people seem to think of it/ask in terms of minutes, so defaulting to
that.

Testing:
Ran a debug build on my phone and observed correct results for some
recent workouts.