Rearrange file hierarchy and move to beekit (#425)
This is a substantial rearrangement of the repository, but mainly
consists of files moves.

There were a large number of files used by each of the different
targets. This moves these files
to the shared BeeKit library so they only need to be built once. This is
less confusing (to me)
and may also help to reduce build times.

For the remaining top level app files also arrange into a number of
subdirectories.

Filesystem structure is also updated to match the XCode structure, and
things in XCode are
alphabetized to reduce the likelihood of merge conflicts in future
changes which introduce files.

Testing:
Confirmed it builds