Store User information in CoreData (#423)
This introduces usage of CoreData, and in particular starts to use it to
store user information. For now, the information is also stored in
preferences to allow undoing the change if needed. This is an
experimental initial change, to evaulate also using CoreData for Goal
and Datapoint information. This should eventually allow simpler data
syncing between the main app and different extensions.

Testing:
Loaded the app and checked login state is persisted
Checked it is possible to log out and in
Check setting default alert start works