Correctly represent deadline in seconds (#434)
In #430 we added the daystamp type. However it was coded to expect
deadline to be a number of minutes, where in fact the beeminder API
returns it as a number of seconds. This lead to incorrectly calculated
start and end times for days, amongst other errors.

Fix this by updating the code and tests to instead expect it to be in
seconds.