Correctly clear Apple Health Integration (#437)
Previously when attempting to disconnect Apple Health goals they would
enter a state where they no longer updated, but beeminder.com still
considered them autodata, and thus e.g. changing the deadline was
forbidden. This was because we were setting the autodata name to an
empty string, rather than null. Sending null was not supported via the
URL encoding schema the app was using, so switch to sending parameters
as JSON instead and send nil.

Fixes #248