Update HealthKit on app launch (#440)
Always update health metrics from apple before syncing with the server,
for example on app launch. Because beeminder often update goals async,
this doesn't guarantee that we will always show correct data reflecting
the latest healthkit data, but should increase the chance. As the app
polls for goals which are recalculating it means it should also show the
correct data shortly after loading.

This might help a little with perceived Apple Health unreliabilty.

Testing:
Launched the app on a real device and observed data