Ignore metadata when suggesting next value (#441)
When choosing the default value to suggest for a new data point, the app
looks at recent data points. Previously it would always choose the most
recent value, but this was not useful if the most recent value was e.g.
a derailment, or a pessimistic presumptive data point. Change the logic
to now ignore these points when finding a value.

Testing:
Manually checked in the app with a goal with relevant data point
Wrote unit tests to extercise the various cases