Show lemniscate for queued goals (#444)
Goals in beeminder can be in a `queued` state. This means while the
client has the latest server data, the server has processing to do to
update the state of the goal. Previously this state was not surfaced in
the app. Here we update the app to show an animated bee over the graph
when in this state. This matches the website behavior.

Some refactoring is done to aid with this. A new component is introduced
to render the lemniscate, and a new image view is introduced to
consistently show the goal image, placeholder, and lemniscate, when
appropriate. Some of the state machines here are somewhat intricate.

Testing:
* Refreshed goals in the app and saw the goal view behaved as expected
* Refreshed goals and the gallery view and observed they behaved as
expected

Known bugs:
* When the gallery view refreshes, the lemniscate can skip between
locations in the animation
* We trigger a graph update, and thus the lemniscate, when refreshing
goals which do not have autodata