Do not hard-crash on missing usernames (#448)
Previously we were using `username!` to look up usernames. This would
throw a fatal exception if the username was not available, crashing the
app. In particular attempts to update HealthKit metrics could hit this
after logging out from the app.

Avoid this by centralizing the username lookup logic in the request
manager making URLs in the app slightly more readable, and have it throw
a catchable exception. This could theoretically cause issues in the case
that a URL contains an additional `{username}` string, but this seems
unlikely.

Testing:
Launched the app on my phone and checked it could load goals.
Signed out and did not immediately experience a crash.