Support webcredentials for easier login (#449)
iOS supports using web credentials to login to apps. It looks like
BeeSwift intended to support this, but it wasn't working properly. The
main fix here was adding a relevant file to the beeminder server, but we
also
* Add the relevant file to this repo, for reference
* Also add www.beeminder.com as a compatible domain

Testing:
Viewed login screen on phone and confirmed password manager credentials
were offered.