Fix calculation for meditation minutes (#453)
Previously meditation minutes used the same algorithm as we used for
time in bed. This was overly generous. It would count minutes containing
meditation, which is different from the apple algorithm which counts
total duration rounded to the nearest minute.

Update the calculation used in the app to match the Apple algorithm.

Fixes #452

Testing:
Put various test data into the simulator, and checked beeminder synced
appropriately.