Remove the "delta text" UI section (#455)
The app previously showed a section of numbers below the graph, e.g. (+1
+1 +2). These were from a time when "lanes" and "road" were part of the
beeminder vocabulary and were pretty hard to understand, especially for
do less goals.

We may wish to replace them with a table of what is due when, but for
now just remove them.

Testing
Loaded the app and looked at a goal.