HealthKit should not remove metadata datapoints (#474)
The HealthKit integration deletes all points which do not correspond to
HealthKit entries. In the past this resulted in deleting points used for
metadata tracking, e.g. derailments or restarts. This was confusing for
support.

Resolve this by ignoring those points when syncing with HealthKit, so
they are neither updated nor deleted.

Fixes #456