some deadlines in timer were not considered
As documented in #17, prior to Commit 5f3278c6f9a2d6fef21d63d7288739d31c81d1d3
the app's built-in Timer did not account for deadlines when submitting datapoints.

While the fix covered many of the possible deadlines, it missed out on the
deadlines extending 'end of day' by up to an hour.

Rather than duplicating the logic to determine the daystamp on the timer,
this fix uses the Daystamp entity introduced in #430, which is meant to
deal with this very thing - given a submission date and a deadline,
calculate which calendar date shall be reported when adding the datapoint.

Fixes #228