handling first hour after midnight deadlines on the goal screen
The add datapoint section of the goal screen contained logic for
determining the stepper's initial value for a goal. It defaults
to today with adjustments made for a goal's deadline.
It contained the same logic as found on the Timer screen and contained
the same logic error of not accounting for deadlines within the first
hour after midnight.

This fix replaces the Goal screen's duplicated logic accounting for
deadlines with logic already present in the dedicated entity Daystamp.

Possibly related: #473