using alamofireimage for transitions when loading thumbnails and graphs
It was already being used in the Today Extension view but not for
the graphs and graph thumbnails elsewhere in the app.
Using it simplifies the code and allows for easily adding transitions.
When scrolling the gallery or loading a goal, the user experience
is enhanced by using smooth rather than jarring animations and transitions.