Use NSModelActor for HealthStoreManager (#519)
This converts HealthStoreManager to be an actor. This allows us to us
NSModelActor and have correct threading for CoreData, hopefully avoiding
some crashes. Because this exposes a method which must be non-async,
there's some usage of `nonisolated`.

Testing:
Ran on device, verified it would sync metrics on startup, and on pull to
refresh.