Convert `CurrentUserManager` to actor (#521)
As part of moving to use `MSModelActor` instead of abitrary background
contexts, this makes `CurrentUserManager` an actor. There remain a few
`nonisolated` methods, though these will largely be removed once the
migration to `CoreData` is complete.

Testing:
Verified that the app launched and could show a dashboard and refresh
goals