initial date in add datapoint looks wrong
Given the scenario, now is 1AM and one is entering into GoalVC for a goal with a deadline of 3AM
the initial date shown was tomorrow. Expected might have been today and better yet is probably
yesterday since 1AM is before the 3AM deadline, thus the daystamp for a datapoint added _now_
is still the (calendar) day before.