Simplify file header template (#534)
This switches to a much shorter template at the top of files. Ideally I
would like to completely remove the header, but that doesn't seem
supported, and results in an empty comment. This at least is shorter
than the default.

Validation:
Created a new file in XCode and checked the comment was as expected.