Color safeSum in forest green as used in the Grayson Dot (#535)
## Summary
The chart includes the Grayson Dot - a different color green applied to
green dots with more than 7d buffer. The iOS app colors the `safeSum`
using colors based on the colors of the dots. This text appears in both
the gallery in a goal's cell as well as in the goal screen itself.
Previously the same green was being used for all `safeSum` with a
`safeBuf` greater than three.

Implements #206

## Validation
Ran app in the simulator using the new colors and logic. Made
screenshots.