Move from StoryBoard to SceneDelegate (#555)
## Summary
Migrate from using a StoryBoard to using SceneDelegate for the main UI
definition.

While almost none of the app was defined via StoryBoard, the top level
still was. This made it harder to open PRs e.g. to change the root view.
As we weren't using the storyboard for anything, instead migrate to
using SceneDelegate to manage this.

## Validation
* [X] Validate the app launches in simulator and shows gallery
* [x] Validate background, relaunch, etc works as before
* [x] Check on iPad the split screen / window behavior has not changed