introduces goal menu with deep links to goal on website (#540)
## Summary
The goal screen contained an action button which opened the goal in an
in-app browser. This way one could manipulate various aspects of the
goal or view data (statistics) from on the website as these features are
lacking in the app.

This merge request replaces the action button on the goal screen, with
single link to the goal page on the website, with a menu with links to
all of the main sections of the website's rendition of the goal:
commitment, stop/pause, data, statistics, and settings. This makes it
easier to more quickly arrive at a particular section. It might also
increase feature discoverability.

For example, the delta text was removed from the app and the current
workaround to check "how much to do to earn x days off" is available on
the website in the statistics section under the "Amounts Due By Day"
subsection.
Also, the app does not support features such as setting the goal's
description or title. It also does not allow editing of datapoints of
goals with autodata whereas the website does.


*For UI changes including screenshots of before and after is great.*
## before
Tapping the action button opens beeminder.com/user/goal in an in-app
browser after which one can navigate through the sections of the goal on
the webpage presented.

## Validation
Opened the app in the simulator. Clicked through various goals and the
corresponding "open this section of the goal on the website" links.